refactor: PropertyDetailDTO 삭제 후, ResponseDTO로 합침.
PropertyDetailDTO의 항목을 PropertyDetailResponseDTO에 합쳐, 하나로 수정.
